refactor: extract AudioObjectID extension from SystemAudioRecorder (#564, #435)
Third slice of #435 (SystemAudioRecorder decomposition). Slices 435a
(aggregate-device identity, #547) and 435b (SystemAudioFileWriter,
#549) shipped earlier.

Moves the `private extension AudioObjectID` block (SystemAudioRecorder.swift
lines 506-674, 169 lines, 5326 bytes, SHA-256 96bf2a8bd5bcc894b59b2647a77eac06136e9f265bc07d138f504a6ba701ad2e)
into a new Sources/BugNarrator/Services/AudioObjectIDExtensions.swift.
The only substantive change is the visibility widen from `private` (file-
scope) to default `internal` (module-scope) — required so that
SystemAudioTapSession (still private in SystemAudioRecorder.swift) can
continue to call the CoreAudio property-read helpers across files.

Method bodies byte-identical (individual method access levels stay at
their current file-scope defaults). Post-widen block (with `private `
stripped from the extension header line) hashes to SHA-256
857f72716faea2fa2fd36aaae5cc154f96bb5b20d7669d96b8b2da824d7d27a2 — the
`private ` prefix drop is the only diff from origin.

Endpoints verified before extraction (`sed -n '506p'` = `private
extension AudioObjectID {`, `'674p'` = `}`). Byte-diff against
origin/main lines 506-674 (with `private ` stripped) is silent (exit 0);
both origin-block SHA and post-widen SHA verified.

The extension contains the CoreAudio property-read helpers: `read<T>` /
`read<T,Q>` with qualifier / `readArray<T>` / `readString`,
`defaultSystemOutputDevice`, `audioDevices`, `deviceUID`,
`audioTapStreamDescription`, and the `systemObject` static. No nested
types. New file imports `CoreAudio` (for AudioObjectID typealias +
AudioObject*/AudioHardware*/kAudio* constants + AudioObjectGetPropertyData*
+ AudioStreamBasicDescription + noErr) and `Foundation` (for CFString
bridging in readString) — verified by codex; AudioToolbox is not
required by this body and AVFoundation is explicitly not needed.

The `private extension DispatchQueue { func drain() async }` block at
SystemAudioRecorder.swift line 496 stays put (out of scope for this
slice per the ticket).

Every existing call site — SystemAudioTapSession's calls to
`AudioObjectID.systemObject.audioDevices()`, `.deviceUID()`,
`.audioTapStreamDescription()`, `.defaultSystemOutputDevice()`, and the
generic-qualifier `.read<T,Q>(...)` at line 487 — resolves cross-file
after the widen. Green 662 with SystemAudioTapSession's CoreAudio path
intact proves it.

Regression only — the CoreAudio helpers cannot be unit-tested without a
real audio session (same constraint that shaped slices 435a/b). 662 unit
tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (4 findings folded:
correct imports CoreAudio+Foundation only, no AVFoundation, missed cross-
file callsite at line 487 covered by the widen, byte-diff protocol
tightened with `private `-prefix strip on origin side to avoid the
"silent except for" contradiction).
